fix u_int8_t seq_num_copy -> it's u_int32_t
git-svn-id: https://svn.mi.fu-berlin.de/des-testbed/Software/DES-SERT/des-aodv/trunk@7738 ec5dfc5b-5239-0410-a031-e37cc464ff32

+++ src/pipeline/aodv_pipeline.c
@@ -287,7 +287,7 @@ int aodv_handle_rreq(dessert_msg_t* msg, size_t len, dessert_msg_proc_t *proc, c
 			// RREQ for me -> answer with RREP
 			dessert_debug("RREQ for me -> answer with RREP to " MAC " over " MAC, EXPLODE_ARRAY6(l25h->ether_shost), EXPLODE_ARRAY6(msg->l2h.ether_shost));
 			pthread_rwlock_wrlock(&pp_rwlock);
-			u_int8_t seq_num_copy = ++seq_num;
+			u_int32_t seq_num_copy = ++seq_num;
 			pthread_rwlock_unlock(&pp_rwlock);
 			dessert_msg_t* rrep_msg = _create_rrep(dessert_l25_defsrc, l25h->ether_shost, msg->l2h.ether_shost, seq_num_copy, AODV_FLAGS_RREP_A);
 			dessert_meshsend_fast(rrep_msg, iface);
